backend/requirements.txt
# ─── Core ────────────────────────────────────────────────────────
Django==4.2.11
djangorestframework==3.15.1

# ─── Base de datos ───────────────────────────────────────────────
psycopg2-binary==2.9.9

# ─── Autenticación JWT ───────────────────────────────────────────
djangorestframework-simplejwt==5.3.1

# ─── CORS ────────────────────────────────────────────────────────
django-cors-headers==4.3.1

# ─── Filtros de búsqueda ─────────────────────────────────────────
django-filter==24.1

# ─── Variables de entorno ────────────────────────────────────────
python-decouple==3.8

# ─── Caché con Redis ─────────────────────────────────────────────
django-redis==5.4.0

# ─── Rate limiting ───────────────────────────────────────────────
django-ratelimit==4.1.0

# ─── Imágenes con Cloudinary ─────────────────────────────────────
cloudinary==1.40.0

# ─── Emails ──────────────────────────────────────────────────────
resend==2.3.0

# ─── Seguridad ───────────────────────────────────────────────────
django-csp==3.8

# ─── Escaneo de vulnerabilidades (CI) ────────────────────────────
pip-audit==2.7.3